bwrap: Report COMMAND exit code in json-status-fd
The exit code is only reported if it exited after a successful exec.

This is accomplished with a pipe, where the write end is closed on exec.

To distinguish between pipe-close pre-exec and at-exec,
data is written to the pipe immediately before calling exec
so if it is closed before exec the pipe is empty
and if it is closed during exec it contains a 1 byte value.

To further distinguish between a successful exec and a failed exec,
on exec failure a second value is written.

Signed-off-by: Richard Maw <richard.maw@codethink.co.uk>

Closes: #257

Closes: #293
Approved by: cgwalters